+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Changed
 - Used rowmapper ineterface
+- Adds email verificaiton queries and tables
 
 ## [1.5.0] - 2020-11-06
 ### Added